% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_glmm.R
\name{bootstrap}
\alias{bootstrap}
\title{Parametric Bootstrap for poisson GLMM}
\usage{
bootstrap(dataset, beta, sigmasq, subject_index = 1, B = 100)
}
\arguments{
\item{dataset}{table: In this dataset, we have one response and one variable indicates groups/subjects the input dataset.}

\item{beta}{vector: the prior value of coefficients used to generate linear predictor, eta.}

\item{sigmasq}{numeric: non-negative. The standard deviation of the random effect. In this project, we assume all the random effect has equal standard deviation.}

\item{subject_index}{integer: a index states which column of dataset is used the first as the group/subject indicator. The default value is subject_index = 1 means column(row) is used as group/subject index.}

\item{B}{integer:positive. The resample size of Bootstrap method. The default size is 100.}
}
\value{
table: a summary table with 8 columns.
}
\description{
Parametric Bootstrap for poisson GLMM
}
\details{
To be convince, let i be the number of groups/subjects, j be the number of
repeated measurement for each subject, B be the number of bootstrap loop (
dafult value is 100)
Conduct bootstrap:
1.Resample k random effects, where k = i * B = (# of subjects/groups * # of
bootstrap iteration).
2.Repeat every generated random effect j times, where j is the number of
repeated measurement.
3.Bind these random effects with intercept and fixed effects provided in
argument 'dataset' to construct design matrix.
4.Calculate linear predictor eta, and hence compute corresponding expected
value of the response variable.
}
\examples{
## provide prior coefficients and sigmasq for generating random effect
full_fit <- run_model(epilepsy,example = "epilepsy")
true_beta <- full_fit$beta
true_sigmasq <- full_fit$sigmasq
## run bootstrap with epilepsy data. We use true_beta for conducting linear
## Run bootstrap with epilepsy data. We use true_beta for conducting linear
## predictors, true_sigmasq for generating random effects. We identify the
## first column as the subject index and set the bootstrap resample size be 100
boot <- bootstrap(dataset = epilepsy, true_beta, true_sigmasq, subject_index = 1, B = 100)
}
