% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_summary.R
\name{BootSummary}
\alias{BootSummary}
\title{Analyse and Summaries bootstrap result}
\usage{
BootSummary(BS_object, SignifLv = 0.05)
}
\arguments{
\item{SignifLv}{numeric: The significance level for conducting confidence interval. The default value is 0.05.}

\item{object}{table: The function return the bootstrap() function in this package. More detail in help(bootstrap)}
}
\value{
table: A 5x5 table. In each row, there is term name, point estimates, bootstrap standard error, Lower Bound and Upper Bound of confidence interval.
}
\description{
Point Estimate, Standard Error and Confidence Interval for bootstrap result
}
\examples{
## provide prior coefficients and sigmasq for generating random effect
full_fit <- run_model(epilepsy,example = "epilepsy")
true_beta <- full_fit$beta
true_sigmasq <- full_fit$sigmasq
## create a bootstrap object, more detail in bootstrap()
boot <- bootstrap(dataset = epilepsy, true_beta, true_sigmasq)
## summaries bootstrap estimates for each term
BS_summary <- BootSummary(boot)
}
